Inventory as template
Changes inventory to a template file.
Adds role "master" to node labels. (Needed to separate nodes when selecting which part of bootstrap script to run)
Sets kubeadm token to default "" so it can be left out of config if Kubernetes is not to be installed. Empty string will not be acceped by KubeAdm so not a security issue.
Generates openstack list of public ip for ansible-inventory with private ip if no public ip is set - to overcome problem with inventory from template